final commit

+++ src/latihanKu/motorClass.java
@@ -9,50 +9,57 @@ package latihanKu;
  * @author ASUS
  */
 public class motorClass {
-    String merek,tipe,warna;
+
+    String merek, tipe, warna;
     int tahun_pembuatan;
-    double harga;
+    int harga;
     boolean status_pajak;
-    
-    public void setmerek(String merekQ){
+
+    public void setmerek(String merekQ) {
         this.merek = merekQ;
     }
-    public  String getmerek(){
+
+    public String getmerek() {
         return this.merek;
     }
-    
-    public void settipe(String tipeQ){
+
+    public void settipe(String tipeQ) {
         this.tipe = tipeQ;
     }
-    public String gettipe(){
+
+    public String gettipe() {
         return this.tipe;
     }
-    
-    public void setwarna(String warnaQ){
+
+    public void setwarna(String warnaQ) {
         this.warna = warnaQ;
     }
-    public String getwarna(){
+
+    public String getwarna() {
         return this.warna;
     }
-    
-    public void settahun_pembuatan(int tahun_pembuatanQ){
+
+    public void settahun_pembuatan(int tahun_pembuatanQ) {
         this.tahun_pembuatan = tahun_pembuatanQ;
     }
-    public int gettahun_pembuatan(){
+
+    public int gettahun_pembuatan() {
         return this.tahun_pembuatan;
     }
-    
-    public void setharga(double hargaQ){
+
+    public void setharga(int hargaQ) {
         this.harga = hargaQ;
     }
-    public double getharga(){
+
+    public int getharga() {
         return this.harga;
     }
-    
-    public void setstatus_pajak(boolean status_pajakQ){
+
+    public void setstatus_pajak(boolean status_pajakQ) {
         this.status_pajak = status_pajakQ;
     }
-    public boolean getstatus_pajak(){
+
+    public boolean getstatus_pajak() {
         return this.status_pajak;
     }
 }
